wtb/database-service/ requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.27.1
psycopg2-binary==2.9.3